Jump big version
The first sprite serves as jump upward and last falling sprite when falling in the three jumps.
The second jump serves as jump forward and the third as backward.
The first sprite in the second row is the position when Gen-an is at the mid-highest location. The second as the highest location (it is cut 'cause the emulator used don't catch what is up the screen). And the third is when he touches the ground.